apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-attendance-api"
version = "0.1.0"
description = "FastAPI backend for face attendance platform"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "numpy>=2.1.0",
  "opencv-python-headless>=4.10.0.84",
  "passlib[argon2]>=1.7.4",
  "pillow>=11.0.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "PyJWT>=2.9.0",
  "python-multipart>=0.0.12",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0"
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
